// src/app/shared/register/model/register.ts
export interface Register {
  linha: string;
  range: string;
  posto: string;
  sum_fimreal_fimplan: number;
  sum_inicioreal_inicioplan: number;
  sum_fimreal_inicioreal: number;
  sum_fimplan_inicioplan: number;
  avg_fimreal_fimplan: number;
  avg_inicioreal_inicioplan: number;
  avg_fimreal_inicioreal: number;
  avg_fimplan_inicioplan: number;
}